requirements.txt
Flask
python-dotenv
requests
gunicorn
flask_session